+++ server/api/v1/sys_user.go
@@ -27,6 +27,17 @@ import (
 func Register(c *gin.Context) {
 	var R request.RegisterStruct
 	_ = c.ShouldBindJSON(&R)
+	UserVerify := utils.Rules{
+		"Username":      {utils.NotEmpty()},
+		"NickName":  {utils.NotEmpty()},
+		"Password":      {utils.NotEmpty()},
+		"AuthorityId":      {utils.NotEmpty()},
+	}
+	UserVerifyErr := utils.Verify(R, UserVerify)
+	if UserVerifyErr!=nil {
+		response.FailWithMessage(UserVerifyErr.Error(), c)
+		return
+	}
 	user := &model.SysUser{Username: R.Username, NickName: R.NickName, Password: R.Password, HeaderImg: R.HeaderImg, AuthorityId: R.AuthorityId}
 	err, userReturn := service.Register(*user)
 	if err != nil {
@@ -45,6 +56,17 @@ func Register(c *gin.Context) {
 func Login(c *gin.Context) {
 	var L request.RegisterAndLoginStruct
 	_ = c.ShouldBindJSON(&L)
+	UserVerify := utils.Rules{
+		"CaptchaId":      {utils.NotEmpty()},
+		"Captcha":  {utils.NotEmpty()},
+		"Username":      {utils.NotEmpty()},
+		"Password":      {utils.NotEmpty()},
+	}
+	UserVerifyErr := utils.Verify(L, UserVerify)
+	if UserVerifyErr!=nil {
+		response.FailWithMessage(UserVerifyErr.Error(), c)
+		return
+	}
 	if captcha.VerifyString(L.CaptchaId, L.Captcha) {
 		U := &model.SysUser{Username: L.Username, Password: L.Password}
 		if err, user := service.Login(U); err != nil {
@@ -134,6 +156,16 @@ func tokenNext(c *gin.Context, user model.SysUser) {
 func ChangePassword(c *gin.Context) {
 	var params request.ChangePasswordStruct
 	_ = c.ShouldBindJSON(&params)
+	UserVerify := utils.Rules{
+		"Username":      {utils.NotEmpty()},
+		"Password":      {utils.NotEmpty()},
+		"NewPassword":   {utils.NotEmpty()},
+	}
+	UserVerifyErr := utils.Verify(params, UserVerify)
+	if UserVerifyErr!=nil {
+		response.FailWithMessage(UserVerifyErr.Error(), c)
+		return
+	}
 	U := &model.SysUser{Username: params.Username, Password: params.Password}
 	if err, _ := service.ChangePassword(U, params.NewPassword); err != nil {
 		response.FailWithMessage("修改失败，请检查用户名密码", c)
@@ -193,6 +225,15 @@ func UploadHeaderImg(c *gin.Context) {
 func GetUserList(c *gin.Context) {
 	var pageInfo request.PageInfo
 	_ = c.ShouldBindJSON(&pageInfo)
+	UserVerify := utils.Rules{
+		"Page": {utils.NotEmpty()},
+		"PageSize": {utils.NotEmpty()},
+	}
+	UserVerifyErr := utils.Verify(pageInfo, UserVerify)
+	if UserVerifyErr!=nil {
+		response.FailWithMessage(UserVerifyErr.Error(), c)
+		return
+	}
 	err, list, total := service.GetUserInfoList(pageInfo)
 	if err != nil {
 		response.FailWithMessage(fmt.Sprintf("获取数据失败，%v", err), c)
@@ -217,6 +258,15 @@ func GetUserList(c *gin.Context) {
 func SetUserAuthority(c *gin.Context) {
 	var sua request.SetUserAuth
 	_ = c.ShouldBindJSON(&sua)
+	UserVerify := utils.Rules{
+		"UUID": {utils.NotEmpty()},
+		"AuthorityId": {utils.NotEmpty()},
+	}
+	UserVerifyErr := utils.Verify(sua, UserVerify)
+	if UserVerifyErr!=nil {
+		response.FailWithMessage(UserVerifyErr.Error(), c)
+		return
+	}
 	err := service.SetUserAuthority(sua.UUID, sua.AuthorityId)
 	if err != nil {
 		response.FailWithMessage(fmt.Sprintf("修改失败，%v", err), c)
@@ -236,6 +286,14 @@ func SetUserAuthority(c *gin.Context) {
 func DeleteUser(c *gin.Context) {
 	var reqId request.GetById
 	_ = c.ShouldBindJSON(&reqId)
+	UserVerify := utils.Rules{
+		"Id": {utils.NotEmpty()},
+	}
+	UserVerifyErr := utils.Verify(reqId, UserVerify)
+	if UserVerifyErr!=nil {
+		response.FailWithMessage(UserVerifyErr.Error(), c)
+		return
+	}
 	err := service.DeleteUser(reqId.Id)
 	if err != nil {
 		response.FailWithMessage(fmt.Sprintf("删除失败，%v", err), c)
